new DatatableHelper and Instance, WIP
